+++ esdoc-lint-plugin/src/Plugin.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const fs = require('fs');
-// hack: using a internal code of esdoc.
-const ASTNodeContainer = require('esdoc/out/src/Util/ASTNodeContainer.js').default;
+// hack: using a internal code of esdocs.
+const ASTNodeContainer = require('esdocs/out/src/Util/ASTNodeContainer.js').default;
 
 /**
  * Lint Output Builder class.