+++ esdoc-publish-html-plugin/test/src/DocumentTest/ExportTest/ArrowFunctionTest.js
@@ -6,13 +6,13 @@ describe('test export arrow function', ()=> {
 
   it('has default import path with direct arrow function definition.', ()=> {
     findParent(doc, '[id="static-function-ArrowFunction"]', '[data-ice="detail"]', (doc)=>{
-      assert.includes(doc, '[data-ice="importPath"]', `import ArrowFunction from 'esdoc-test-fixture/src/Export/ArrowFunction.js'`);
+      assert.includes(doc, '[data-ice="importPath"]', `import ArrowFunction from 'esdocs-test-fixture/src/Export/ArrowFunction.js'`);
     });
   });
 
   it('has named import path with direct arrow function definition.', ()=>{
     findParent(doc, '[id="static-function-testExportArrowFunction2"]', '[data-ice="detail"]', (doc)=>{
-      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction2} from 'esdoc-test-fixture/src/Export/ArrowFunction.js'`);
+      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction2} from 'esdocs-test-fixture/src/Export/ArrowFunction.js'`);
     });
   });
 
@@ -27,13 +27,13 @@ describe('test export arrow function', ()=> {
 
   it('has named import path with undocument', ()=>{
     findParent(doc, '[id="static-function-testExportArrowFunction4"]', '[data-ice="detail"]', (doc)=>{
-      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction4} from 'esdoc-test-fixture/src/Export/ArrowFunction.js'`);
+      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction4} from 'esdocs-test-fixture/src/Export/ArrowFunction.js'`);
     });
   });
 
   it('has named import path with indirect function definition.', ()=> {
     findParent(doc, '[id="static-function-testExportArrowFunction5"]', '[data-ice="detail"]', (doc)=>{
-      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction5} from 'esdoc-test-fixture/src/Export/ArrowFunction.js'`);
+      assert.includes(doc, '[data-ice="importPath"]', `import {testExportArrowFunction5} from 'esdocs-test-fixture/src/Export/ArrowFunction.js'`);
     });
   });
 });