+++ esdoc-publish-html-plugin/test/src/DocumentTest/ExportTest/DefaultTest.js
@@ -7,6 +7,6 @@ import {readDoc, assert} from './../../util.js';
 describe('TestExportDefault', ()=> {
   const doc = readDoc('class/src/Export/Default.js~TestExportDefault.html');
   it('has default import path.', ()=> {
-    assert.includes(doc, '.header-notice [data-ice="importPath"]', `import TestExportDefault from 'esdoc-test-fixture/src/Export/Default.js'`);
+    assert.includes(doc, '.header-notice [data-ice="importPath"]', `import TestExportDefault from 'esdocs-test-fixture/src/Export/Default.js'`);
   });
 });