+++ esdoc-publish-html-plugin/test/src/DocumentTest/ExportTest/VariableIndirectDefaultTest.js
@@ -6,7 +6,7 @@ describe('test export variable indirect default', ()=> {
 
   it('has default import path with indirect variable definition.', ()=> {
     findParent(doc, '[id="static-variable-testExportVariableIndirectDefault"]', '[data-ice="detail"]', (doc)=>{
-      assert.includes(doc, '[data-ice="importPath"]', `import testExportVariableIndirectDefault from 'esdoc-test-fixture/src/Export/VariableIndirectDefault.js'`);
+      assert.includes(doc, '[data-ice="importPath"]', `import testExportVariableIndirectDefault from 'esdocs-test-fixture/src/Export/VariableIndirectDefault.js'`);
     });
   });
 });