+++ esdoc-standard-plugin/src/Plugin.js
@@ -2,17 +2,17 @@ class Plugin {
   onHandlePlugins(ev) {
     const option = ev.data.option || {};
     const plugins = [
-      {name: 'esdoc-lint-plugin', option: option.lint},
-      {name: 'esdoc-coverage-plugin', option: option.coverage},
-      {name: 'esdoc-accessor-plugin', option: option.accessor},
-      {name: 'esdoc-type-inference-plugin', option: option.typeInference},
-      {name: 'esdoc-external-ecmascript-plugin'},
-      {name: 'esdoc-brand-plugin', option: option.brand},
-      {name: 'esdoc-undocumented-identifier-plugin', option: option.undocumentIdentifier},
-      {name: 'esdoc-unexported-identifier-plugin', option: option.unexportedIdentifier},
-      {name: 'esdoc-integrate-manual-plugin', option: option.manual},
-      {name: 'esdoc-integrate-test-plugin', option: option.test},
-      {name: 'esdoc-publish-html-plugin'}
+      {name: 'esdocs-lint-plugin', option: option.lint},
+      {name: 'esdocs-coverage-plugin', option: option.coverage},
+      {name: 'esdocs-accessor-plugin', option: option.accessor},
+      {name: 'esdocs-type-inference-plugin', option: option.typeInference},
+      {name: 'esdocs-external-ecmascript-plugin'},
+      {name: 'esdocs-brand-plugin', option: option.brand},
+      {name: 'esdocs-undocumented-identifier-plugin', option: option.undocumentIdentifier},
+      {name: 'esdocs-unexported-identifier-plugin', option: option.unexportedIdentifier},
+      {name: 'esdocs-integrate-manual-plugin', option: option.manual},
+      {name: 'esdocs-integrate-test-plugin', option: option.test},
+      {name: 'esdocs-publish-html-plugin'}
     ];
 
     const existPluginNames = ev.data.plugins.map(plugin => plugin.name);