+++ script/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 
-for plugin in $(ls -1 -d esdoc-*)
+for plugin in $(ls -1 -d esdocs-*)
 do
   echo $plugin
   (cd $plugin && npm run test)