+++ esdoc-publish-markdown-plugin/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## Next
 - **Feat**
-  - Add `option.filename` ([#56](https://github.com/esdoc/esdoc-plugins/pull/56)). Thanks [@Verikon](https://github.com/Verikon)
+  - Add `option.filename` ([#56](https://github.com/esdocs/esdocs-plugins/pull/56)). Thanks [@Verikon](https://github.com/Verikon)