+++ esdoc-react-plugin/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 (2017-09-10)
 - **Fix**
-  - Crash when `esdoc-publish-html-plugin`@1.1.0
+  - Crash when `esdocs-publish-html-plugin`@1.1.0
 
 ## 1.0.0 (2017-07-30)
 - Release